Elevation check files
These are the files used each day during the reef construction project to check the progress of the rock elevations. What would happen is the surveys would work at night when the tides were low to survey the rocks, and then Bill would receive either while in the field with the surveyors or the next day the rock elevation information.  The code is annotated to make a variety of figures including the box plots of elevation by reef that people seem to like.